requirements-thawed.txt
gunicorn
# gunicorn[gevent]
gevent
greenlet
redis
requests
backoff
boto3
Fabric3
# executor
# envoy
BaseHash
PyJWT
pydash
PyYAML>=5.4
toolz
xxhash
lxml
pandas
enasearch
attrdict
robox
python-magic
urllib3
Jinja2
fs  # PyFilesystem
sentry-sdk
humanize
typing-extensions  # backport, not required after Python 3.8
text-unidecode
pymmh3
webdav4>=0.10.0
ptvsd

# Temporary version of paramiko that merges PR (OpenSSH key format): https://github.com/paramiko/paramiko/pull/1343
# Alternative workaround - generate PEM format keys like:
#   ssh-keygen -t rsa -b 4096 -f ~/.ssh/my-compute-key -m PEM
# Install with PARAMIKO_REPLACE=1 environment variable set
# https://github.com/ploxiln/paramiko-ng/archive/2.8.0.tar.gz#egg=paramiko-ng
# cryptography>=2.5
paramiko

# Base Django packages
django
django-reversion
django-cors-headers>=4.9.0,<5.0
django-extensions
ipython
django-storages
django-object-actions
django-cache-memoize
social-auth-core
social-auth-app-django>=5.4.0
rest-social-auth
django-rest-framework-social-oauth2==1.2.0
django-oauth-toolkit==3.1.0
django-guardian
djangorestframework-guardian
djangorestframework-csv
# whitenoise
# whitenoise[brotli]

# for Django django.contrib.auth.hashers
bcrypt
argon2-cffi

# django-jsonfield, the schinckel one, now at https://github.com/adamchainz/django-jsonfield
git+https://github.com/adamchainz/django-jsonfield#egg=django-jsonfield

# The 'other' json field library
# jsonfield  # https://github.com/dmkoch/django-jsonfield
# compatibility layer for jsonfield<->Django native Postgres JSONField
# django-jsonfield-compat  # broken ?

# For LoginRequiredMixin, CsrfExemptMixin
django-braces
django-environ==0.12.0
# django-tus

# Authentication
# Is django-allauth actually used ?
# django-allauth
# django-auth-adfs
# https://github.com/NIF-au/django-rapid-connect-demo

# Django REST Framework, and some optional extra packages
djangorestframework
markdown
django-filter
# django-filter
django-guardian
coreapi
djangorestframework-simplejwt
drf-spectacular

# drf_openapi==1.3.0
# git+https://github.com/limdauto/drf_openapi.git#egg=drf_openapi
# https://github.com/limdauto/drf_openapi/issues/105
# git+https://github.com/maykinmedia/drf_openapi.git@105-proxy-label-cannot-be-decoded#egg=drf_openapi  # Removed

git+https://github.com/pansapiens/json-merge-patch.git#egg=json_merge_patch
jsonpatch
jsonschema

# Celery task queue
celery
flower
django-celery-results
django-celery-beat
setproctitle

# Postgres packages
# sudo apt-get install postgresql postgresql-server-dev-all
psycopg2-binary
inflection>=0.3.1
